Rename present poll to session and update imports
