+++ frontend/src/components/present-session/present-footer/present-footer.tsx
