+++ frontend/src/components/present-session/present-preview/present-preview.scss
