+++ frontend/src/components/present-session/present.scss
